requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
requests==2.32.3
drf-spectacular==0.27.2